% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_warn_or_error_ticket.R
\name{write_warn_or_error_ticket}
\alias{write_warn_or_error_ticket}
\title{Return and write any system errors and warnings for an expression to the working directory README}
\usage{
write_warn_or_error_ticket(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
If there are errors and warnings to report, they are returned in the console as well as in the working directory README.md.
}
\description{
This function is meant to catch errors to trace back to the steps in a workflow where things bottlenecked.
}
